requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cupy-cuda12x==13.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.0
fastrlock==0.8.3
filelock==3.17.0
flatbuffers==25.2.10
fonttools==4.55.8
fsspec==2025.2.0
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
ipython_pygments_lexers==1.1.1
jax==0.5.2
jaxlib==0.5.1
jedi==0.19.2
Jinja2==3.1.5
kiwisolver==1.4.8
lap==0.5.12
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
portalocker==3.1.1
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0.8
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
-e git+ssh://git@github.com/rasmusmade/thesisrepo.git@3dc73f83fbf02e29c16712037f1c69a2577e659e#egg=SAM_2
scipy==1.15.1
screeninfo==0.8.1
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sounddevice==0.5.1
stack-data==0.6.3
supervision==0.25.1
sympy==1.13.1
tokenizers==0.21.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.2
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.86
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
